pyproject.toml
[project]
name = "fastapi-mongo-modern-devcontainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=1.26.0",
    "fastapi>=0.114.0",
    "pydantic-settings>=2.4.0",
    "uvicorn>=0.30.6",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "pre-commit>=3.8.0",
    "ruff>=0.6.4",
]